+++ front/errors.rb
@@ -12,47 +12,40 @@
 #
 # THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFINGEMENT. IN NO EVENT SHALL THE
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 # AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-require 'minitest/autorun'
-require 'webmock/minitest'
-require_relative 'test__helper'
-require_relative '../objects/wts'
-require_relative '../objects/btc'
+not_found do
+  status 404
+  content_type 'text/html', charset: 'utf-8'
+  haml :not_found, layout: :layout, locals: merged(
+    page_title: 'Page not found'
+  )
+end
 
-class WTS::BtcTest < Minitest::Test
-  # Fake BTC
-  class FakeBtc
-    def initialize(addr)
-      @addr = addr
+error do
+  e = env['sinatra.error']
+  if e.is_a?(WTS::UserError)
+    settings.log.error("#{request.url}: #{e.message}")
+    body(Backtrace.new(e).to_s)
+    headers['X-Zold-Error'] = e.message[0..256]
+    if params[:noredirect]
+      content_type 'text/plain', charset: 'utf-8'
+      return Backtrace.new(e).to_s
     end
-
-    def create
-      { hash: @addr, pvt: 'empty' }
-    end
+    flash('/', e.message, error: true)
   end
-
-  def test_validates_txn
-    btc = WTS::Btc.new(log: test_log)
-    assert(btc.trustable?(27_900, 6))
-  end
-
-  def test_monitors_txns
-    skip
-    WebMock.allow_net_connect!
-    btc = WTS::Btc.new(log: test_log)
-    assets = WTS::Assets.new(WTS::Pgsql::TEST.start, log: test_log)
-    addr = '1BPs843gTEkfNk9LL6Lr2QyRNmWtQvJejv'
-    assets.acquire('johnny-l09')
-    btc.monitor(assets) do |hash, txn, satoshi, confirmations|
-      assert_equal(addr, hash)
-      assert_equal('154a7dfd574abbc5d22ebca1b8ca9358dcf545e84eaffb46c4978b981af9c13e', txn)
-      assert_equal(25_800, satoshi)
-      assert(confirmations > 100)
-    end
+  status 503
+  Raven.capture_exception(e, extra: { 'request_url' => request.url })
+  if params[:noredirect]
+    Backtrace.new(e).to_s
+  else
+    haml :error, layout: :layout, locals: merged(
+      page_title: 'Error',
+      error: Backtrace.new(e).to_s
+    )
   end
 end
